requirements.txt
mcp>=0.1.0
typing-extensions>=4.0.0
uvicorn>=0.15.0
starlette>=0.25.0 
pydantic>=2.6.0
httpx>=0.24.0
PyYAML>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
